splite the profile in acomponent to reuseable in dashboard #31
