+++ client/src/components/pages/application/Profile/index.js
@@ -1,15 +1,10 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import Checkbox from '@material-ui/core/Checkbox';
-import { useHistory } from 'react-router-dom';
 import { Helmet } from 'react-helmet';
-import Card from '../../../common/card';
 import Limitation from '../../../common/limitation';
-import Typography from '../../../common/Typography';
-import Button from '../../../common/Button';
+import Profile from './profile';
 import backGround from '../../../../assets/images/backGround.svg';
 import profileImage from '../../../../assets/images/HomeImg.svg';
-
 import './index.css';
 
 const applicantData = async (id) => {
@@ -34,11 +29,10 @@ const getUserID = async () => {
   return data;
 };
 
-const Profile = () => {
+const ProfilePage = () => {
   const [data, setData] = useState();
   const [tec, setTec] = useState();
   const [UserId, setId] = useState();
-  const history = useHistory();
 
   useEffect(() => {
     if (!UserId)
@@ -66,260 +60,8 @@ const Profile = () => {
         className="profileImage profile"
       />
 
-      {data ? (
-        <div className="Profile__container">
-          <Card
-            ClassName="card_profile"
-            content={
-              <table className="Profile__table">
-                <tr>
-                  <td colSpan="2">
-                    <img
-                      src={data.avatar}
-                      alt={data.fullName}
-                      className="Profile__avatar"
-                    />
-                  </td>
-                </tr>
-                <tr className="doted">
-                  <td>
-                    <Typography variant="body2" color="primary" align="left">
-                      User Name
-                    </Typography>
-                  </td>
-                  <td>
-                    <Typography variant="body2" align="left">
-                      {data.fullName}
-                    </Typography>
-                  </td>
-                </tr>
-                <tr className="doted">
-                  <td>
-                    <Typography variant="body2" color="primary" align="left">
-                      Availability
-                    </Typography>
-                  </td>
-                  <td />
-                </tr>
-                <tr>
-                  <td className="item">
-                    <Typography variant="body2" align="left">
-                      Email
-                    </Typography>
-                  </td>
-                  <td>
-                    <Typography variant="body2" align="left">
-                      {data.email}
-                    </Typography>
-                  </td>
-                </tr>
-                <tr>
-                  <td className="item">
-                    <Typography variant="body2" align="left">
-                      Gender
-                    </Typography>
-                  </td>
-                  <td>
-                    <Typography variant="body2" align="left">
-                      {data.gender}
-                    </Typography>
-                  </td>
-                </tr>
-                <tr className="doted">
-                  <td className="item">
-                    <Typography variant="body2" align="left">
-                      Age
-                    </Typography>
-                  </td>
-                  <td>
-                    <Typography variant="body2" align="left">
-                      {data.age}
-                    </Typography>
-                  </td>
-                </tr>
-                <tr className="doted">
-                  <td>
-                    <Typography variant="body2" color="primary" align="left">
-                      Accounts
-                    </Typography>
-                  </td>
-                  <td />
-                </tr>
-                <tr>
-                  <td className="item">
-                    <Typography variant="body2" align="left">
-                      Github
-                    </Typography>
-                  </td>
-                  <td>
-                    <Typography variant="body2" align="left">
-                      {data.githubLink}
-                    </Typography>
-                  </td>
-                </tr>
-                <tr>
-                  <td className="item">
-                    <Typography variant="body2" align="left">
-                      FreeCode Camp
-                    </Typography>
-                  </td>
-                  <td>
-                    <Typography variant="body2" align="left">
-                      {data.freeCodeCampLink}
-                    </Typography>
-                  </td>
-                </tr>
-                <tr className="doted">
-                  <td className="item">
-                    <Typography variant="body2" align="left">
-                      Code Wares
-                    </Typography>
-                  </td>
-                  <td>
-                    <Typography variant="body2" align="left">
-                      {data.codeWarsLink}/
-                    </Typography>
-                  </td>
-                </tr>
-                <tr className="doted">
-                  <td>
-                    <Typography variant="body2" color="primary" align="left">
-                      Technical Task
-                    </Typography>
-                  </td>
-                  <td />
-                </tr>
-                {tec &&
-                  tec.map(({ taskName }) => (
-                    <tr>
-                      <td className="item" colSpan="2">
-                        <Typography variant="body2" align="left">
-                          <Checkbox
-                            checked={data && data.technicalTasks}
-                            className="profile_check_box"
-                            color="primary"
-                          />
-                          {taskName}
-                        </Typography>
-                      </td>
-                    </tr>
-                  ))}
-                <tr className="doted">
-                  <td>
-                    <Typography variant="body2" color="primary" align="left">
-                      Project
-                    </Typography>
-                  </td>
-                  <td />
-                </tr>
-                <tr>
-                  <td className="item" colSpan="2">
-                    <Typography variant="body2" align="left">
-                      <Checkbox
-                        checked={data && data.projectGithubLink}
-                        className="profile_check_box"
-                        color="primary"
-                      />
-                      <a
-                        className="Link"
-                        href={data.projectGithubLink}
-                        target="popup"
-                        onClick={() => {
-                          window.open(
-                            'https://discord.com/',
-                            'popup',
-                            'width=600,height=600,scrollbars=no,resizable=no'
-                          );
-                        }}
-                      >
-                        Project Github Link
-                      </a>
-                    </Typography>
-                  </td>
-                </tr>
-                <tr className="doted">
-                  <td>
-                    <Typography variant="body2" color="primary" align="left">
-                      Submitted
-                    </Typography>
-                  </td>
-                  <td />
-                </tr>
-                <tr>
-                  <td className="item">
-                    <Typography variant="body2" align="left">
-                      Code Wares Score
-                    </Typography>
-                  </td>
-                  <td>
-                    <Typography variant="body2" color="primary" align="left">
-                      {data.codeWarsKyu}
-                    </Typography>
-                  </td>
-                </tr>
-                <tr>
-                  <td className="item">
-                    <Typography variant="body2" align="left">
-                      FreeCode Camp Score
-                    </Typography>
-                  </td>
-                  <td>
-                    <Typography variant="body2" color="primary" align="left">
-                      {data.freeCodeCampPoints}
-                    </Typography>
-                  </td>
-                </tr>
-                <tr>
-                  <td className="item" colSpan="2">
-                    <Typography variant="body2" align="left">
-                      <Checkbox
-                        checked={data && data.freeCodeCampTopics}
-                        className="profile_check_box"
-                        color="primary"
-                      />
-                      FreeCode Camp Topics
-                    </Typography>
-                  </td>
-                </tr>
-                <tr>
-                  <td className="item" colSpan="2">
-                    <Typography variant="body2" align="left">
-                      <Checkbox
-                        checked={data && data.applicationSubmittedDate}
-                        className="profile_check_box"
-                        color="primary"
-                      />
-                      Submitted
-                      {`on ${
-                        data.applicationSubmittedDate &&
-                        data.applicationSubmittedDate.split('T')[0]
-                      }`}
-                    </Typography>
-                  </td>
-                </tr>
-                <tr className="doted">
-                  <td className="item" colSpan="2">
-                    <Typography variant="body2" align="left">
-                      <Checkbox
-                        checked={data && data.accepted}
-                        className="profile_check_box"
-                        color="primary"
-                      />
-                      Accepted
-                    </Typography>
-                  </td>
-                </tr>
-                <tr>
-                  <td colSpan="2" className="profile__btn">
-                    <Button onClick={() => history.push('/steps')}>
-                      Go Back
-                    </Button>
-                  </td>
-                </tr>
-              </table>
-            }
-          />
-        </div>
+      {data && tec ? (
+        <Profile data={data} tec={tec} next="/steps" />
       ) : (
         <div className="loading">
           <Limitation />
@@ -329,4 +71,4 @@ const Profile = () => {
   );
 };
 
-export default Profile;
+export default ProfilePage;
